requirements.txt
# 数据采集
google-play-scraper>=1.2.2
app-store-scraper>=0.3.5
playwright>=1.40.0  # 用于App Store采集的替代方案（更稳定）

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 文本分析
jieba>=0.42.1
snownlp>=0.12.3
textblob>=0.17.1
nltk>=3.8.1
vaderSentiment>=3.3.2

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.2

# 工具库
pyyaml>=6.0
requests>=2.23.0  # 注意：app-store-scraper需要requests==2.23.0，这里使用>=以兼容
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 其他
tqdm>=4.65.0
python-dateutil>=2.8.2

